pyproject.toml
[project]
name = "pybites-pdc-snipster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jwattier", email = "jasonwattier@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pre-commit>=4.3.0",
    "sqlmodel>=0.0.27",
    "typing>=3.10.0.0",
]

[project.scripts]
pybites-pdc-snipster = "pybites_pdc_snipster:main"

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest-cov>=7.0.0",
]
